Done some change on User permissions

+++ food/views.py
@@ -4,6 +4,10 @@ from .models import Item
 from django.template import loader
 from .forms import ItemForm
 from django.core.paginator import Paginator
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
+
 def index(request):
     item_list= Item.objects.all()
     #context ={ 
@@ -27,27 +31,37 @@ def detail(request,id):
     }    
     return render(request,'food/detail.html',context)
 
+@login_required
 def create_item(request):
     form=ItemForm(request.POST or None)
 
     if form.is_valid():
-        form.save()
+        item=form.save(commit=False)
+        item.user_name=request.user
+        item.save()
         return redirect('food:index')
     
     return render(request,'food/item-form.html',{'form':form})
 
+@login_required
 def update_item(request,id):
     item=Item.objects.get(id=id)
-    form=ItemForm(request.POST or None, instance=item)
-    if form.is_valid():
-        form.save()
-        return redirect('food:index')
+    if item.user_name != request.user:
+        return HttpResponseForbidden("You dont have the permission to edit this item,")
+    if request.method=='POST':
+        form=ItemForm(request.POST or None, instance=item)
+        if form.is_valid():
+            form.save()
+            return redirect('food:index')
     
     return render(request,'food/item-form.html',{'form':form,'item':item})
 
+@login_required
 def delete_item(request,id):
     item=Item.objects.get(id=id)
-    
+    if item.user_name != request.user:
+       return HttpResponseForbidden("You dont have the permission to delete this item,")
+     
     if request.method=='POST':
         item.delete()
         return redirect('food:index')
